Moved functions away from index.js in frog-utils
