+++ .dependency-cruiser.json
@@ -1,0 +1,36 @@
+{
+  "forbidden": [
+    {
+      "name": "no-circular",
+      "severity": "error",
+      "from": {},
+      "to": {
+        "circular": true
+      }
+    },
+    {
+      "name": "no-orphans",
+      "severity": "warn",
+      "from": {
+        "orphan": true,
+        "pathNot": "\\.d\\.ts$"
+      },
+      "to": {}
+    }
+  ],
+  "options": {
+    "doNotFollow": {
+      "dependencyTypes": [
+        "npm",
+        "npm-dev",
+        "npm-optional",
+        "npm-peer",
+        "npm-bundled",
+        "npm-no-pkg"
+      ]
+    },
+    "exclude": {
+      "path": "__tests__"
+    }
+  }
+}
