+++ frog-utils/src/A.js
@@ -1,0 +1,16 @@
+// @flow
+
+import * as React from 'react';
+
+export const A = ({ onClick, children, ...rest }: any): any => (
+  <a
+    href="#"
+    onClick={e => {
+      e.preventDefault();
+      onClick();
+    }}
+    {...rest}
+  >
+    {children}
+  </a>
+);
