+++ frog-utils/src/WikiContext.js
@@ -1,0 +1,6 @@
+// @flow
+
+import * as React from 'react';
+
+// $FlowFixMe need to specify a context type
+export const WikiContext = React.createContext([]);
