+++ frog-utils/src/booleanize.js
@@ -1,0 +1,3 @@
+// @flow
+
+export const booleanize = (bool: string): boolean => bool === 'true';
