+++ frog-utils/src/cloneDeep.js
@@ -1,0 +1,27 @@
+// @flow
+
+export const cloneDeep = (o: any): any => {
+  let newO;
+  let i;
+
+  if (typeof o !== 'object') return o;
+
+  if (!o) return o;
+  if (o instanceof Date) return new Date(o.valueOf());
+  if (Object.prototype.toString.apply(o) === '[object Array]') {
+    newO = [];
+    for (i = 0; i < o.length; i += 1) {
+      newO[i] = cloneDeep(o[i]);
+    }
+    return newO;
+  }
+
+  newO = {};
+  // eslint-disable-next-line no-restricted-syntax
+  for (i in o) {
+    if (Object.prototype.hasOwnProperty.call(o, i)) {
+      newO[i] = cloneDeep(o[i]);
+    }
+  }
+  return newO;
+};
