+++ frog-utils/src/currentDate.js
@@ -1,0 +1,6 @@
+// @flow
+
+export const currentDate = (): string => {
+  const d = new Date();
+  return d.toString();
+};
