+++ frog-utils/src/flattenOne.js
@@ -1,0 +1,7 @@
+// @flow
+
+export const flattenOne = (ary: any[]): any[] =>
+  ary.reduce(
+    (acc: any[], x: any) => (Array.isArray(x) ? [...acc, ...x] : [...acc, x]),
+    []
+  );
