+++ frog-utils/src/getDisplayName.js
@@ -1,0 +1,11 @@
+// @flow
+
+export const getDisplayName = (WrappedComponent: any): string => {
+  if (typeof WrappedComponent.displayName === 'string') {
+    return WrappedComponent.displayName;
+  } else if (typeof WrappedComponent.name === 'string') {
+    return WrappedComponent.name;
+  } else {
+    return 'Component';
+  }
+};
