+++ frog-utils/src/getInitialState.js
@@ -1,0 +1,12 @@
+// @flow
+
+export const getInitialState = (activities: Object, d: number = 1) => {
+  const n = Math.floor(activities.length / 2);
+  return n === 0
+    ? activities[0]
+    : {
+        direction: d > 0 ? 'row' : 'column',
+        first: getInitialState(activities.slice(0, n), -d),
+        second: getInitialState(activities.slice(n, activities.length), -d)
+      };
+};
