+++ frog-utils/src/getSlug.js
@@ -1,0 +1,9 @@
+// @flow
+
+import { shuffle } from 'lodash';
+
+const groupchars = 'ABCDEFGHIJKLMNOPQRSTUWXYZ123456789'.split('');
+export const getSlug = (n: number) =>
+  shuffle(groupchars)
+    .slice(0, n)
+    .join('');
