+++ frog-utils/src/isBrowser.js
@@ -1,0 +1,11 @@
+// @flow
+
+const doesWindowObjExist = () => {
+  try {
+    return window !== undefined;
+  } catch (e) {
+    return false;
+  }
+};
+
+export const isBrowser = doesWindowObjExist();
