+++ frog-utils/src/notEmpty.js
@@ -1,0 +1,9 @@
+// @flow
+
+// checks that some of the values in an object are not empty
+export const notEmpty = (obj: Object): boolean => {
+  if (!obj) {
+    return false;
+  }
+  return Object.keys(obj).reduce((acc, val) => acc || Boolean(obj[val]), false);
+};
