+++ frog-utils/src/splitAt.js
@@ -1,0 +1,8 @@
+// @flow
+
+// list utils
+export const splitAt = (i: number, xs: Array<any>): Array<Array<any>> => {
+  const first = xs.slice(0, i);
+  const second = xs.slice(i, xs.length);
+  return [first, second];
+};
