+++ frog-utils/src/withVisibility.js
@@ -1,0 +1,16 @@
+// @flow
+
+import { compose, withHandlers, withState } from 'recompose';
+
+export const withVisibility: Function = compose(
+  withState('visible', 'setVisibility', false),
+  withHandlers({
+    toggleVisibility: ({ setVisibility }) => x => {
+      if (typeof x === 'boolean') {
+        setVisibility(x);
+      } else {
+        setVisibility(n => !n);
+      }
+    }
+  })
+);
