+++ frog-utils/src/wordWrap.js
@@ -1,0 +1,20 @@
+// @flow
+
+export const wordWrap = (text: string, maxLength: number): string[] => {
+  const result = [];
+  let line = [];
+  let length = 0;
+  text.split(' ').forEach(word => {
+    if (length + word.length >= maxLength) {
+      result.push(line.join(' '));
+      line = [];
+      length = 0;
+    }
+    length += word.length + 1;
+    line.push(word);
+  });
+  if (line.length > 0) {
+    result.push(line.join(' '));
+  }
+  return result;
+};
