+++ frog-utils/src/zipList.js
@@ -1,0 +1,4 @@
+// @flow
+
+export const zipList = (xs: Array<any>): Array<any> =>
+  xs[0].map((_, i) => xs.map(x => x[i]));
