+++ package.json
@@ -38,6 +38,7 @@
     "babel-plugin-captains-log": "^1.1.0",
     "chokidar": "^2.1.0",
     "cypress": "^3.3.1",
+    "dependency-cruiser": "^4.25.0",
     "enzyme": "^3.8.0",
     "eslint": "^6.0.0",
     "eslint-config-airbnb": "^17.1.0",
